Remove Tagset checks in ModifiedDistanceExt
GetTags() is expensive because it creates a new tagSet and then merges in another tagSet into it before its returned. Contains on a tagSet is also relatively expensive.

There is already a stat version of the ImmobileUnit tag that exists and is checked for in this same method. when this stat is initialized, it now checks for the tag on the unit to correctly set the stats default value, removing the need to check for the tag later.

This improves AI performance.